add wav reader

+++ port_audio_stream_parameter.cpp
@@ -53,7 +53,7 @@ void PortAudioStreamParameter::_bind_methods() {
 	ADD_PROPERTY(PropertyInfo(Variant::INT, "device_index"), "set_device_index", "get_device_index");
 	ADD_PROPERTY(PropertyInfo(Variant::INT, "channel_count"), "set_channel_count", "get_channel_count");
 	ADD_PROPERTY(PropertyInfo(Variant::INT, "sample_format", PROPERTY_HINT_ENUM, "FLOAT_32, INT_32, INT_24, INT_16, INT_8, U_INT_8, CUSTOM_FORMAT, NON_INTERLEAVED"), "set_sample_format", "get_sample_format");
-	ADD_PROPERTY(PropertyInfo(Variant::INT, "suggested_latency"), "set_suggested_latency", "get_suggested_latency");
+	ADD_PROPERTY(PropertyInfo(Variant::REAL, "suggested_latency"), "set_suggested_latency", "get_suggested_latency");
 
 	// PortAudioSampleSize
 	BIND_ENUM_CONSTANT(FLOAT_32);
